requirements.txt
# Core dependencies
pikepdf>=9.0.0
pillow>=10.0.0
numpy>=1.24.0
pymupdf>=1.23.0

# Optional for quality metrics
scikit-image>=0.21.0

# Development
pytest>=7.0.0
black>=23.0.0
pylint>=2.17.0
